requirements.txt
google-genai
python-dotenv
ipywidgets
pillow